small fixes and updates
-updated service form's icons and icon mouse hover image indexes to work
like google drive. left them disabled for now to match the disabled grey
scale in switcher form until implemented
-handled UI to inform user there is an update on both the main service
form and the switcher service form. had to make the variable that holds
the latest value global (not really any room left on the main form but
if we want to put it there to we can make it bigger)
-changed version check for entire string instead of major and minor
since that did not work correctly when a minor update was available
-changed label1 to InfoLabel on the main form
-other small visual changes to service form
-removed old comment code in serviceform and program.cs as they aren't
really needed but thanks for the info